migraciones corregidas de la base de datos cambio del tipo de dato de las llaves foraneas

+++ database/migrations/2023_10_03_185235_create_empleado_turnos_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('empleado_turnos', function (Blueprint $table) {
             $table->id();
-            $table->integer('turno_id');
-            $table->integer('empleado_id');
+            $table->unsignedBigInteger('turno_id');
+            $table->unsignedBigInteger('empleado_id');
             $table->date('fecha');
             $table->time('horaEntrada');
             $table->time('horaLlegada');
